Improved range, timestamps and output file. -Fixed variable range across frames. Currently fixed manually. -Implemented output file as mp4 instead of gif. -Added timestamp to frames.

+++ Animate.py
@@ -7,44 +7,54 @@ import os
 
 print(datetime.now())           	###############################################
 
-
+# Initialization
 WRFfiles=[]
 PNGfiles=[]
 
-
+cleanpng=1
 dir_path="/mnt/seaes01-data01/dmg/dmg/mbessdl2/Spanish_Plume/WRF/run-zrek/"
+range_min=980
+range_max=1040
+variable="slp"
+ptitle="Sea level pressure [hPa]"
+outfile="SLP"
 
-#Get list of files from directoy
+# Get list of files from directoy
 for file in os.listdir(dir_path):
-    # check only wrfout files
     if file.startswith('wrfout'):
         WRFfiles.append(file)
 WRFfiles.sort()
 
-#Plots each time slice in each file
+# Plot each time frame in each file
 for wrf_fn in WRFfiles:
 	# Open the NetCDF file
 	print("Loading ",wrf_fn)
 	ncfile = Dataset(dir_path+wrf_fn)
 
-	# Get number of time slices and plot them
+	# Get number of time frames and plot them
 	timerange=ncfile.variables['Times'].shape[0]
-#	if timerange>3:timerange=3
+#	if timerange>3:timerange=1                              ## For tests only
 	for ti in range(timerange):
-		of=wrf_fn+"_t_"+str(ti)+".png"
+		of=outfile+wrf_fn+"_t_"+str(ti)+".png"
 		PNGfiles.append(of)
 		print("Processing:",ti+1,"/",timerange, end = '\r')
-		Plot2DField(ncfile,"slp","Sea level pressure [hPa] - t="+str(ti),ti,of)
+		Plot2DField(ncfile,variable,ptitle,range_min,range_max,ti,of)
 	print("Processed successfully.")
 
 # Build GIF
-with imageio.get_writer('SLP.gif', mode='I') as writer:
+#with imageio.get_writer(outfile+".gif", mode='I') as writer:
+#    for filename in PNGfiles:
+#        image = imageio.imread(filename)
+#        writer.append_data(image)
+# Build mp4
+with imageio.get_writer(outfile+".mp4", mode='I') as writer:
     for filename in PNGfiles:
         image = imageio.imread(filename)
         writer.append_data(image)
 		
-# Remove individual frames
-for file in PNGfiles:
-	os.remove(file)
+# Remove individual frame files
+if cleanpng:
+    for file in PNGfiles:
+	    os.remove(file)
 
 print(datetime.now())           	###############################################
